G H F float 으로 바꿈 station 클래스 GHF 0 기본값 초기화, parentId -> parentName

+++ AStarStationProject/Assets/02.Scripts/AStar.cs
@@ -9,9 +9,9 @@ public class AStar : MonoBehaviour
 
     private GameObject rings;
     private int nowLine;
-    private int G;
-    private int F;
-    private int H;
+    private float G;
+    private float F;
+    private float H;
     private List<Station> openedList = new List<Station>();
     private List<Station> closedList = new List<Station>();
     private Station nowStation;
